Development

+++ react-app/src/components/ProductDetailsPage/index.js
@@ -18,6 +18,7 @@ import UpdateReviewModal from '../UpdateReviewModal/UpdateReviewModal';
 
 import './ProductDetails.css';
 import { useModal } from '../../context/Modal';
+import LoginFormModal from '../LoginFormModal';
 
 function ProductDetailsPage() {
 
@@ -25,12 +26,11 @@ function ProductDetailsPage() {
     const [ quantity, setQuantity ] = useState(1);
     const [ currentImageIndex, setCurrentImageIndex ] = useState(0);
     const [ cartMessage, setCartMessage ] = useState("");
-    const [ error, setError ] = useState(null);
+    const [ error] = useState(null);
     const { openModal } = useModal()
     const { productId } = useParams();
     const dispatch = useDispatch();
 
-    const cartItems = useSelector(state => state.cartitems.currentCart);
     const product = useSelector(state => state.product.productDetails);
     const favorites = useSelector(state => state.favorite.favorites);
     const userId = useSelector(state => {
@@ -40,7 +40,6 @@ function ProductDetailsPage() {
           return state.session.user && state.session.user.id;
         }
       });
-    console.log(userId)
     const reviews = useSelector(state => state.review.reviews)
 
 
@@ -65,9 +64,13 @@ function ProductDetailsPage() {
         event.stopPropagation();
         event.preventDefault();
 
+        if (!userId) {
+            openModal(<LoginFormModal />);
+            return;
+        };
+
         if (isFavorited) {
             const favoriteToRemove = favorites.find((favorite) => favorite.productId === product.id);
-            console.log('favoritetoRemove in ProductDetailsPage --------------------', favoriteToRemove)
             if (favoriteToRemove) {
                 dispatch(removeFavoriteThunk(favoriteToRemove.userId, favoriteToRemove.productId));
             }
@@ -77,7 +80,6 @@ function ProductDetailsPage() {
     };
 
     const handleAddToCart = () => {
-        const existingItem = cartItems.find(item => item.productId === product.id);
 
         dispatch(addToCartThunk({
             productId: product.id,
